refactor: FetchedEntryとスキーマをSharedに移動
FetchedEntryとそのスキーマは一連の変更でPostEntryやEditEntryからも参照されるようになったため共有空間に置いたほうがいいと判断.
