+++ UpHateblo.Lib.Tests/Entry/Post/PostableEntryOperatorTests.cs
@@ -1,6 +1,6 @@
 using UpHateblo.Lib.Entry.Edit;
-using UpHateblo.Lib.Entry.List;
 using UpHateblo.Lib.Entry.Post;
+using UpHateblo.Lib.Entry.Shared;
 
 namespace UpHateblo.Lib.Tests.Entry.Post;
