+++ UpHateblo.Lib/Entry/Edit/EditableEntry.cs
@@ -1,8 +1,8 @@
 using System.ComponentModel.DataAnnotations;
 using Equatable.Attributes;
-using UpHateblo.Lib.Entry.List;
 using UpHateblo.Lib.Entry.Post;
 using UpHateblo.Lib.Entry.Read;
+using UpHateblo.Lib.Entry.Shared;
 
 namespace UpHateblo.Lib.Entry.Edit;
