+++ UpHateblo.Lib/Entry/Post/PostableEntry.cs
@@ -1,8 +1,8 @@
 using System.ComponentModel.DataAnnotations;
 using Equatable.Attributes;
 using UpHateblo.Lib.Entry.Edit;
-using UpHateblo.Lib.Entry.List;
 using UpHateblo.Lib.Entry.Read;
+using UpHateblo.Lib.Entry.Shared;
 
 namespace UpHateblo.Lib.Entry.Post;
